Move exploration helpers to steps/
